Refactoring records to file namespace
