+++ Records/Earnings/Box.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace NumaxLab\Icaa\Records;
+namespace NumaxLab\Icaa\Records\Earnings;
 
 use Assert\Assert;
 use Assert\Assertion;
 use Carbon\Carbon;
+use NumaxLab\Icaa\Records\RecordInterface;
 use Stringy\Stringy;
 
 class Box implements RecordInterface
