+++ Records/Earnings/CinemaTheatre.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace NumaxLab\Icaa\Records;
+namespace NumaxLab\Icaa\Records\Earnings;
 
 use Assert\Assert;
 use Assert\Assertion;
+use NumaxLab\Icaa\Records\RecordInterface;
 use Stringy\Stringy;
 
 class CinemaTheatre implements RecordInterface
