+++ Tests/Records/Earnings/CinemaTheatreTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace NumaxLab\Icaa\Tests\Records;
+namespace NumaxLab\Icaa\Tests\Records\Earnings;
 
-use NumaxLab\Icaa\Records\CinemaTheatre;
+use NumaxLab\Icaa\Records\Earnings\CinemaTheatre;
 use NumaxLab\Icaa\Records\RecordInterface;
 use PHPUnit\Framework\TestCase;
 use Stringy\Stringy;
